requirements.txt
aiogram==3.4.1
aiohttp==3.9.3
pydantic==2.6.1
requests==2.31.0
dnspython==2.4.2
python-dotenv==1.0.0
pydantic==2.4.2
